pyproject.toml
[tool.poetry]
name = "snowflake-ml"
version = "0.0.0-dev"  # set dynamically
description = ""
authors = ["Murilo Cunha <murilo@dataroots.io>"]

[tool.poetry.dependencies]
python = "~3.8"
snowflake-snowpark-python = {extras = ["pandas"], version = "^0.8.0"}
scikit-learn = "1.1.1"
joblib = "^1.1.0"
scipy = "1.7.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
streamlit = "^1.11.1"
htbuilder = "^0.6.0"
toml = "^0.10.2"
mistune = "^2.0"
matplotlib = "^3.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Devtools not in environment, but used in `pre-commit`
[tool.mypy]
disallow_untyped_calls = false
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "snowflake_ml._vendor.*"
disallow_untyped_defs = false

[tool.isort]
profile = "black"

[tool.sqlfluff.core]
dialect = "snowflake"

[tool.sqlfluff.rules.L010] # keywords
capitalisation_policy = "upper"

[tool.sqlfluff.rules.L016] # max-length
ignore_comment_clauses = true
ignore_comment_lines = true
tab_space_sizes = 2
max_line_length = 90
